Suggest valid attribute names during project validation
This commit improves the error messages produced during project schema
validation so that, where possible, we suggest valid attribute names
that the user may have meant to type. For instance, if they provide a
"Name" attribute where we wanted "name" (lowercase "n"), we'll now say
so. Where there is not a close match, we'll enumerate the full list of
valid names to try and guide the user.

Matching is implemented using Levenshtein distances and ignores case.
Some examples of the new functionality:

* `{"Name": ...}` yields `project is missing a 'name' attribute; found
  'Name' instead`
* `{..., "rutnime": ...}` yields `project is missing a 'runtime'
  attribute; found 'rutnime' instead`
* `{..., "template": {"displayNameDisplayName": ...}, ...}` yields
  `'displayNameDisplayName' not allowed; the allowed attributes are
  'config', 'description', 'displayName', 'important', 'metadata' and
  'quickstart'`