Fix --continue-on-error running indefinitely when a resource fails to be created or updated (#16371)
<!--- 
Thanks so much for your contribution! If this is your first time
contributing, please ensure that you have read the
[CONTRIBUTING](https://github.com/pulumi/pulumi/blob/master/CONTRIBUTING.md)
documentation.
-->

# Description

This PR ensures that we only run the `StepCompleteFunc` if the step is
successful. If not, we should use the step's `Fail` method to register
the result. Previously, when a user runs a pulumi up operation with
`--continue-on-error`, the engine would register the result twice, once
in
https://github.com/pulumi/pulumi/blob/ee6ec150d8237d472196da9e87265b6b0f24e6af/pkg/resource/deploy/step_executor.go#L342
and if a failure occurs, it is re-registered in
https://github.com/pulumi/pulumi/blob/ee6ec150d8237d472196da9e87265b6b0f24e6af/pkg/resource/deploy/step_executor.go#L373
I suspect that the double registration is the cause of the user facing
error. Since this would only occur for create and update steps, it also
explains why the error doesn't occur during deletions.

Testing:

- Manually ran the repro identified in #16373 and
https://github.com/pulumi/pulumi-command/issues/435#issuecomment-2114250614
against a locally built CLI with this fix.
- Extended existing lifecycle tests to account for the different code
branches of `CreateStep.Apply()` and `UpdateStep.Apply()`

Fixes #16373

## Checklist

- [x] I have run `make tidy` to update any new dependencies
- [x] I have run `make lint` to verify my code passes the lint check
  - [x] I have formatted my code using `gofumpt`

<!--- Please provide details if the checkbox below is to be left
unchecked. -->
- [x] I have added tests that prove my fix is effective or that my
feature works
<!--- 
User-facing changes require a CHANGELOG entry.
-->
- [x] I have run `make changelog` and committed the
`changelog/pending/<file>` documenting my change
<!--
If the change(s) in this PR is a modification of an existing call to the
Pulumi Cloud,
then the service should honor older versions of the CLI where this
change would not exist.
You must then bump the API version in
/pkg/backend/httpstate/client/api.go, as well as add
it to the service.
-->
- [ ] Yes, there are changes in this PR that warrants bumping the Pulumi
Cloud API version
<!-- @Pulumi employees: If yes, you must submit corresponding changes in
the service repo. -->