Remove `Step.Apply`'s `preview` parameter (#16367)
With the clean up of deployment options (#16357), it is no longer
necessary for `Step.Apply` methods to accept a `preview bool` parameter.
This information is already present on the receiver's `deployment`
property and removing the argument a. allows us to consolidate the
surface area of the API, hopefully preventing bugs and b. further
improve consistency around the use of "preview" (referring to the
`preview` action) and "dry run" (referring to any part of an action that
does not perform modifications, such as showing what will happen during
a `refresh` or `destroy`).