resolve providers in the engine before running transforms
We're considering whether it makes sense to document setting "default
providers" through transforms.  For that we want to make sure the
transforms behave exactly the way we expect before trying to document
anything, and going ahead with that plan.

Currently it is the SDKs responsibility to resolve the providers map
and to resolve the parent relationship and send the provider in a
resolved manner to the engine.  For example in the NodeJS SDK this
currently happens here: https://github.com/pulumi/pulumi/blob/91568bc03bcdc4564f242887390e33848dddd26e/sdk/nodejs/resource.ts#L445-L469

To make sure this works in the same way in every SDK, and that we can
add an engine test to make sure this works consistently, we add the
same resolution of the provider before we pass these options to the
transform.  We'll keep the resolution in the SDK, which will be
preferred (we'll choose the Provider resource option first in any
case), but now also have a fallback to the same extend in the engine.

Also add a test to show the behaviour is correct to the engine.