Fix a panic when ignoring wildcard values with arrays of different length (#16406)
The panic occurred during pulumi preview while calculating values for
ignored changes with a `*` wildcard, when compared arrays had different
length. We weren't checking the array boundaries carefully, and thus
panicing when going out of bounds.

After the fix, the same repro in the issue leads to a preview error

> cannot ignore changes to the following properties because one or more
elements of the path are missing:
"defaultActions[*].forward.targetGroups[*].weight"

This seems consistent with other code branches and tests, where `reset`
returns false for non-matching shapes. User's expectation may be that we
succeed in this case - curious to get a take from ones who implemented
the wildcard feature in the past.

Fixes #16403