Bump the minimum version of `pulumi` that generated Python SDKs depend on
When we generate Python provider SDKs and programs, we generate
dependencies on the core Pulumi Python SDK (which lives in
`sdk/python`). When we add new features to the core Python SDK, such as
the `@deprecated` decorator added in #16400, we need to bump the minimum
version of the core SDK that generated code will depend on. If we don't,
we risk generating code that doesn't work, or whose dependency graphs
can't be resolved. This commit bumps the version for the latest set of
changes made to the SDK, so that generated dependencies will reference
at least version 3.121.1 of the `pulumi` Python package.