Parallelize Construct and Call calls in the Node.js SDK
Due to global state issues in the nodejs runtime, we serialized MLC construct requests in #6452.
This had the downside that cloud resources were not created in parallel if multiple instances
of the same component were instantiated in the same program.

This change removes the serialization of Construct and Call calls.
The global state issues were already fixed in #10568