fix: advise people of the ESLint Plugin for catching interpolation issues (#15854)
Raised by @lunaris in onboarding, we don't advertise the [Pulumi ESLint
Plugin](https://github.com/pulumi/eslint-plugin-pulumi). As using an
output in a template literal is one of the most common avenues to
discovering this issue, perhaps we should advertise the plugin here?

What's lacking from this is documentation, but I imagine that users will
see this, create issues, and that will drive improvements either in the
base SDK or the ESLint Plugin.