add RegisterResourceTransform as alias for RegisterStackTransform (#16435)
In subsequent PRs we're going to introduce transforms for Calls, Invokes
and Reads. Add an alias to the current register stack transforms
functions, and deprecate the old version. This is both for consistency,
and because it's a bit confusing that register stack transforms only
works for register resource calls, and not other types of pulumi
operations.

/xref https://github.com/pulumi/pulumi/issues/16171