fix race condition when registering stack transform (#16443)
When registering a stack transform we want to make sure it is finished
before any of the subsequent resource registrations are being run. We
try to do this with a `.then`/`.finally` construct. However the callback
that's passed to the `registerStackTransform` function is called
asynchronously, and thus the `.finally` may run before the stack
transform is actually finished.

Fix this by creating a new promise, and only resolving/rejecting it when
the callback has finished running.

Fixes https://github.com/pulumi/pulumi/issues/15639