Set the --continue-on-error flag with PULUMI_CONTINUE_ON_ERROR environment variable (#16442)
If the `PULUMI_CONTINUE_ON_ERROR` environment variable is set to a
truthy value, it will have the effect of `--continue-on-error` for `up`
and `destroy` commands. This helps streamline setting the flag,
including through Automation API and Deployments.

Part of #16405 but only for `--continue-on-error` that was specifically
requested by a customer. If this lands, we can apply the same pattern
elsewhere.