Update `upload-artifact` and `download-artifact` actions
v2 of `actions/upload-artifact` and `actions/download-artifact` will be
deprecated on June 30, 2024. This commit bumps the versions of these
actions to v4. The main breaking change between v2/v3 and v4 is that
artifacts are now immutable, so multiple uploads to artifacts with the
same name will fail by default. To resolve this we introduce explicit
names for all previously-implicitly-named artifacts, and modify the
hashes we compute for test runs to include whether or not the test is an
integration test or not (since we have tests with the same name in both
suites e.g. `pkg`, `sdk`).