Wrap call&construct of node provider server in AsyncLocalStorage
In #16428 we parallelized the call&construct calls of the node
provider server. It was wrongly assumed that the state of the
settings is localized to the current call, but that is only
done when running in `asyncLocalStorage.run`.
This change wraps the call&construct calls in `asyncLocalStorage.run`
to correctly localize state.