implement skeleton command for `pulumi state move`
Implement a skeleton command for `pulumi state move`.  This can so far
only more a single resource from one stack to another, including
copying the provider it needs, and leave valid state files behind.

These state files are not written to the backend yet, nor is there any
kind of UI.  Further PRs will build on top of this.

The command is intentionally commented out so this can be merged
independently.