fudge numbers to avoid flakyness in token renewal tests (#16495)
The current numbers are too tight on the busy CI runners, where
sometimes running the test itself is just a bit slow. Try to fudge the
numbers a bit to make this test less likely to flake. This won't
completely eliminate the problem, but hopefully make it *much* less
likely.