make sure ci-ok is red even if there are skipped tests (#16507)
When there are skipped tests, any jobs depending on the skipped test
will also be skipped. This means ci-ok will be skipped in that case. For
some reason GitHub believes this should mean "Go ahead ignoring this
check" instead of treating it the same as a failed check.

Therefore when a test is skipped in the CI script because an earlier
test failed, ci-ok is marked as Skipped and we still enqueue the PR,
wasting time and resources.

The workaround is to make sure this check *always* runs, and checking
the result of the scripts it depends on. We do the same in the on-merge
job.