reduce the usaage of PULUMI_BOT_TOKEN
PULUMI_BOT_TOKEN is a shared resource across the company and we should
be using it as little as possible.  This will hopefully help us bump
into the API rate limit less.

Do this using a combination of
- not installing stuff we don't seem to need
- using `secrets.GITHUB_TOKEN` instead.  `secrets.GITHUB_TOKEN` is
scoped per repository, so any excessive usage is solely on us to fix.