rewrite the URN when resources are being moved between projects
When a resource is moved between stacks in different projects we also
need to rewrite project part of the URNs.  Do that here.  Moving
between different projects already works by virtue of the requireStack
function supporting that when providing the fully qualified name of
the stack.

There is some awkwardness here as old diy backends may not return a
project name, in which case we error out here.