Wrap call&construct of node provider server in AsyncLocalStorage (#16487)
In #16428 we parallelized the call&construct calls of the node provider
server. It was wrongly assumed that the state of the settings are
localized to the current call, but that is only done when running in
`asyncLocalStorage.run`.
This change wraps the call&construct calls in `asyncLocalStorage.run` to
correctly localize state.

Addresses
https://github.com/pulumi/pulumi/pull/16428#issuecomment-2190932525