Update `upload-artifact` and `download-artifact` actions (#16485)
v2 of `actions/upload-artifact` and `actions/download-artifact` will be
deprecated on June 30, 2024. This commit bumps the versions of these
actions to v4. The main breaking change between v2/v3 and v4 is that
artifacts are now immutable, so multiple uploads to artifacts with the
same name will fail by default. In this case there are two options --
use different names, and recollect the artifacts later with a modified
`download-artifact` instantiation, or pass `overwrite: true` to calls
which will happen later to first delete the artifact and create a new
one in its place. It _appears_ that we only fall foul of repeated
uploads in one place, and that for this case we can use `overwrite`, so
that is the approach this commit takes.

---------

Co-authored-by: Julien Poissonnier <julien@caffeine.lu>
Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>