Display an actionable error message when dependency installation fails (#16489)
When the dependency installation fails because we could not find the
packagemanager executable, tell the user how to remedy the error and
finish the dependency installation.

Example:
```
The packagemangager to use for installing dependencies pnpm
Installing dependencies...

error: installing dependencies failed: Could not find `pnpm` executable.
Install pnpm from https://pnpm.io/installation and make sure it is in your PATH.
Run `pulumi install` to complete the installation.
```