only retry tests when they are running in the merge queue
We recently tried to remove the retry script from this repo.  However
there were too many flaky tests, which made merging anything very
difficult.  This was however a very useful exercise to remove some
flaky tests, which would be useful to continue doing.

To compromise on that, run tests with retries in the merge queue, but
don't do retries in other situations, such as PRs.  It's easier to
retry tests manually there anyway, but flaky tests are still going to
be more likely to be noticed.